4789: Move ProjectStepDuplication to Timeline:StepDuplication
